Project/hello-cdk/requirements-dev.txt
pytest==6.2.5
attrs==21.4.0
aws-cdk-lib==2.12.0
cattrs==1.10.0
cdk-ec2-key-pair==3.2.0
cdk-iam-floyd==0.300.0
constructs==10.0.58
jsii==1.53.0
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
typing_extensions==4.0.1